Improves the structuring of the project
